requirements.txt
pyramid==2.0
psycopg2-binary==2.9.10